+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ### Features
 
+ - Add ability to use Actions in Custom Functions
  - Add Input Action: Execute Python 3 Code ([#1334](https://github.com/kizniche/Mycodo/issues/1334))
+ - Add Function: Adafruit Neokey (Key Press Executes Actions) ([#1353](https://github.com/kizniche/Mycodo/issues/1353))
+ - Change deprecated threading.currentThread to threading.current_thread
 
 ### Bugfixes